+++ readable_infer_type.d.ts
@@ -1,2 +1,2 @@
 import type { Readable } from './readable';
-export declare type readable_infer_type<I extends Readable<unknown>> = I extends (infer O) ? O : never;
+export declare type readable_infer_type<I> = I extends Readable<infer O> ? O : never;