+++ readable_infer_type.ts
@@ -1,3 +1,5 @@
 import type { Readable } from './readable'
-export type readable_infer_type<I extends Readable<unknown>> =
-	I extends (infer O) ? O : never
+export type readable_infer_type<I> =
+	I extends Readable<infer O>
+		? O
+		: never